Code Clean Up
