+++ MercadoPagoSDK/MercadoPagoSDK/String+Additions.swift
@@ -85,39 +85,3 @@ extension String {
         return NSMutableAttributedString(string: self, attributes: attributes)
     }
 }
-
-
-private class Localizator {
-    
-    static let sharedInstance = Localizator()
-    
-    lazy var localizableDictionary: NSDictionary! = {
-        var bundle: Bundle? = MercadoPago.getBundle()
-        if bundle == nil {
-            bundle = Bundle.main
-        }
-        let languageBundle = Bundle(path : MercadoPagoContext.getLocalizedPath())
-        let languageID = MercadoPagoContext.getLocalizedID()
-        
-        if let path = languageBundle?.path(forResource: "Localizable_\(languageID)", ofType: "plist") {
-            return NSDictionary(contentsOfFile: path)
-        }
-        fatalError("Localizable file NOT found")
-    }()
-    
-    func localize(string: String) -> String {
-        guard let localizedStringDictionary = localizableDictionary.value(forKey: string) as? NSDictionary, let localizedString = localizedStringDictionary.value(forKey: "value") as? String else {
-            assertionFailure("Missing translation for: \(string)")
-            return ""
-        }
-
-        return localizedString
-    }
-}
-
-extension String {
-    var localized_beta: String {
-        return Localizator.sharedInstance.localize(string: self)
-    }
-}
-
